update: FavouriteCharactersListScreen
Previously, the CharacterViewModel function getFavouriteCharacters would
be invoked within a LaunchedEffect scope in this screen. This
implementation was ineffective since the function would be invoked
everytime the screen enters composition. Currently, the
favouriteCharacters state has been exposed as a StateFlow using the
stateIn operator and onStart method that invokes the
getFavouriteCharacters method when there is an active subscriber